feat: 更新日志配置，调整日志级别和格式，添加错误处理模块

+++ src/entity/devices.rs
@@ -5,8 +5,8 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(table_name = "devices")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
-    pub user_id: i32,
+    pub id: i64,
+    pub user_id: i64,
     pub token: String,
     pub create_time: DateTime,
     pub update_time: DateTime,
